CoBot/requirements.txt
fastapi
uvicorn
llama-index-core
llama-index-embeddings-huggingface
llama-index-llms-nvidia
python-dotenv
chromadb-client
llama-index-vector-stores-chroma
llama-index-readers-mongodb
pymongo
docling
python-multipart

